add obe process interface and get obe process api for lecturer

+++ course-service/course-service/src/main/java/hcmuaf/edu/vn/fit/course_service/dto/response/TeacherCourseResponse.java
@@ -16,7 +16,7 @@ public class TeacherCourseResponse {
     private String courseCode;
     private String courseName;
     private String courseTitle;
-    private long studentCount;
+    private int studentCount;
     private String semester;
     private int obeProgress;
     private String lecturerName;
